paginated_items: revert to previous mechanism, letting inventory_browser setting up its pagination
to prevent breaking app/modules/welcome/lib/show_paginated_items.js
