+++ app/modules/inventory/components/paginated_items.svelte
@@ -3,34 +3,16 @@
   import viewport from '#lib/components/actions/viewport'
   import { wait } from '#lib/promises'
   import { onChange } from '#lib/svelte/svelte'
-  import { clone } from 'underscore'
 
-  export let Component, itemsIds, componentProps
-
-  const allowMore = true
+  export let Component, componentProps, pagination
 
   let items = [], flash, waiting
 
-  let hasMore, fetchMore
-  async function setupPagination () {
-    items = []
-    const remainingItems = clone(itemsIds)
-    hasMore = () => remainingItems.length > 0
-    fetchMore = async () => {
-      const batch = remainingItems.splice(0, 20)
-      if (batch.length === 0) return
-      await app.request('items:getByIds', { ids: batch, items })
-      items = items
-    }
-    fetch()
-  }
-
   function fetch () {
     waiting = fetchMore()
+      .then(() => items = pagination.items)
       .catch(err => flash = err)
   }
-
-  $: onChange(itemsIds, setupPagination)
 
   let bottomElInView = false
   async function bottomIsInViewport () {
@@ -48,6 +30,15 @@
   function bottomLeftViewport () {
     bottomElInView = false
   }
+
+  let fetchMore, hasMore, allowMore
+  function reinitializePagination () {
+    if (!pagination) return
+    ;({ fetchMore, hasMore, allowMore } = pagination)
+    fetch()
+  }
+
+  $: onChange(pagination, reinitializePagination)
 </script>
 
 <div class="paginated-items">
